idling resources
